% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert documents}
\usage{
convert(file_path, save_path = ".", method = c("auto", "bibr", "grobid"), ...)
}
\arguments{
\item{file_path}{Path to the document file, or a directory of documents}

\item{save_path}{Path to a directory in which to save the JSON file}

\item{method}{whether to use bibr or grobid to convert a file}

\item{...}{further arguments to pass to convert_bibr or convert_grobid}
}
\value{
the path to the JSON file
}
\description{
Uses grobid or bibr to convert a PDF file to paper format
}
