% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.batch_query}
\alias{.batch_query}
\title{Batch query}
\usage{
.batch_query(
  urls,
  batch_size = 5,
  msg = "Batch Query",
  delay = 0.5,
  accept = "application/json",
  req_func = function(req) {
     req
 }
)
}
\arguments{
\item{urls}{A vector of URLs}

\item{batch_size}{Size of each batch}

\item{msg}{Message to show in progress bar - set to NULL to omit progressbar}

\item{delay}{Courtesy delay between batches (in seconds)}

\item{accept}{The header type to accept}
}
\value{
a list of responses
}
\description{
Batch query
}
\keyword{internal}
