% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{fig_image_view}
\alias{fig_image_view}
\title{View a figure image}
\usage{
fig_image_view(paper, figure_id = 1)
}
\arguments{
\item{paper}{a paper object}

\item{figure_id}{the id for the figure to show}
}
\value{
plots the figure
}
\description{
View a figure image
}
\examples{
paper <- demopaper()
fig_image_view(paper, 1)
fig_image_view(paper, 2)
}
