% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_info}
\alias{osf_info}
\title{Retrieve info from the OSF by ID}
\usage{
osf_info(osf_ids, pb = NULL)
}
\arguments{
\item{pb}{a progress bar passed from another function}

\item{osf_id}{an OSF ID or URL}
}
\value{
a data frame of information
}
\description{
Retrieve info from the OSF by ID
}
\keyword{internal}
