% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_node_data}
\alias{osf_node_data}
\title{#' Retrieve info from the OSF by ID
#'
#' @param osf_id an OSF ID or URL
#' @param pb a progress bar passed from another function
#'
#' @returns a data frame of information
#' @export
#' @keywords internal
osf_info <- function(osf_id,
Structure OSF Node Data}
\usage{
osf_node_data(data)
}
\arguments{
\item{data}{the data object from an OSF API call}
}
\value{
a data frame with a subset of data
}
\description{
#' Retrieve info from the OSF by ID
#'
#' @param osf_id an OSF ID or URL
#' @param pb a progress bar passed from another function
#'
#' @returns a data frame of information
#' @export
#' @keywords internal
osf_info <- function(osf_id,
Structure OSF Node Data
}
\keyword{internal}
