% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{read}
\alias{read}
\title{Read in grobid XML or bibr JSON}
\usage{
read(file_path, include_images = FALSE)
}
\arguments{
\item{file_path}{path to a directory containing XML and/or JSON files, or a vector of paths}

\item{include_images}{whether to include images in the figures table of the paper object (they make object size larger, only relevant to bibr imports)}
}
\value{
a paper or paperlist
}
\description{
Read in grobid XML or bibr JSON
}
