% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{validate_paper}
\alias{validate_paper}
\title{Validate a Paper Object}
\usage{
validate_paper(paper)
}
\arguments{
\item{paper}{a paper object}
}
\value{
TRUE or error
}
\description{
Checks if a paper object conforms to the JSON schema.
}
\examples{
paper <- list(paper_id = "Not a paper object")
validate_paper(paper)

paper <- demopaper()
validate_paper(paper)
}
