% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{read_bibr}
\alias{read_bibr}
\title{Read Bibr zip file}
\usage{
read_bibr(file_path)
}
\arguments{
\item{file_path}{path to the zip file}
}
\value{
a paper object
}
\description{
Read Bibr zip file
}
\keyword{internal}
